PPF-29 Setup laravel nova

+++ app/Domain/Subscriptions/Subscription.php
@@ -15,9 +15,9 @@ final class Subscription
         public readonly string $description,
         public readonly IntegrationType $integrationType,
         public readonly Currency $currency,
-        public readonly int $price,
+        public readonly float $price,
         public readonly BillingInterval $billingInterval,
-        public readonly ?int $fee,
+        public readonly ?float $fee,
     ) {
     }
 }
